update readme and more cleanup
