+++ src/components/atoms/HtmlFragment.tsx
@@ -1,0 +1,7 @@
+export interface HtmlFragmentProps {
+  html: string
+}
+
+export const HtmlFragment: React.FC<HtmlFragmentProps> = ({ html }) => {
+  return <div dangerouslySetInnerHTML={{ __html: html }} />
+}
