+++ src/pages/posts/index.tsx
@@ -1,5 +1,5 @@
 import { GetStaticProps, NextPage } from 'next'
-import { PostList } from '~/components/molecules/PostList'
+import { PostList } from '~/components/organisms/PostList'
 import { BaseTemplate } from '~/components/templates/BaseTemplate'
 import { HomepageTemplate } from '~/components/templates/HomepageTemplate'
 import { getAllPosts } from '~/util/getAllPosts'
@@ -14,16 +14,7 @@ const Posts: NextPage<{ posts: PostMarkdown[]; aboutMe: MarkdownRaw }> = ({
       meta={{
         title: 'post index'
       }}
-      Content={
-        <HomepageTemplate
-          Content={
-            <div className='my-10'>
-              <h1>Posts</h1>
-              <PostList posts={posts} />
-            </div>
-          }
-        />
-      }
+      Content={<HomepageTemplate Content={<PostList posts={posts} />} />}
     />
   )
 }
